Move isDateInRange to lib/utils from team and timesheet
