+++ frontend/packages/app/src/app/pages/timesheet/index.tsx
@@ -29,7 +29,7 @@ import AddTime from "@/app/components/AddTime";
 import { TimesheetTable } from "@/app/components/timesheet-table";
 import { SubmitButton } from "@/app/components/timesheet-table/components/submitButton";
 import { Header, Main } from "@/app/layout/root";
-import { parseFrappeErrorMsg, expectatedHours, copyToClipboard } from "@/lib/utils";
+import { parseFrappeErrorMsg, expectatedHours, copyToClipboard, isDateInRange } from "@/lib/utils";
 import { RootState } from "@/store";
 import { WorkingFrequency } from "@/types";
 import { HolidayProp, LeaveProps, NewTimesheetProps, timesheet } from "@/types/timesheet";
@@ -37,7 +37,7 @@ import { Approval } from "./components/approval";
 import { EditTime } from "./components/editTime";
 import ExpandableHours from "./components/expandableHours";
 import { initialState, reducer } from "./reducer";
-import { isDateInRange, validateDate } from "./utils";
+import { validateDate } from "./utils";
 import { InfiniteScroll } from "../resource_management/components/infiniteScroll";
 
 function Timesheet() {
